Bug#29853758: COMPILATION ERRORS WITH GCC 9.1.1
Problem: New signed/unsigned warnings are treated as errors.

Solution: Remove unnecessary tests/casts.

+++ testclients/mysql_client_test.cc
@@ -5545,7 +5545,7 @@ static void bind_date_conv(uint row_count, bool preserveFractions) {
   my_bind[2].buffer_type = MYSQL_TYPE_DATETIME;
   my_bind[3].buffer_type = MYSQL_TYPE_DATE;
 
-  for (i = 0; i < (int)array_elements(my_bind); i++) {
+  for (i = 0; i < array_elements(my_bind); i++) {
     my_bind[i].buffer = (void *)&tm[i];
     my_bind[i].is_null = &is_null[i];
     my_bind[i].length = &length[i];
@@ -5567,7 +5567,7 @@ static void bind_date_conv(uint row_count, bool preserveFractions) {
   check_execute(stmt, rc);
 
   for (count = 0; count < row_count; count++) {
-    for (i = 0; i < (int)array_elements(my_bind); i++) {
+    for (i = 0; i < array_elements(my_bind); i++) {
       tm[i].neg = 0;
       tm[i].second_part = second_part + count;
       if (my_bind[i].buffer_type != MYSQL_TYPE_TIME) {
@@ -16775,7 +16775,7 @@ static void test_bug40365(void) {
   memset(my_bind, 0, sizeof(my_bind));
   my_bind[0].buffer_type = MYSQL_TYPE_DATETIME;
   my_bind[1].buffer_type = MYSQL_TYPE_DATE;
-  for (i = 0; i < (int)array_elements(my_bind); i++) {
+  for (i = 0; i < array_elements(my_bind); i++) {
     my_bind[i].buffer = (void *)&tm[i];
     my_bind[i].is_null = &is_null[i];
   }
@@ -16783,7 +16783,7 @@ static void test_bug40365(void) {
   rc = mysql_stmt_bind_param(stmt, my_bind);
   check_execute(stmt, rc);
 
-  for (i = 0; i < (int)array_elements(my_bind); i++) {
+  for (i = 0; i < array_elements(my_bind); i++) {
     tm[i].neg = 0;
     tm[i].second_part = 0;
     tm[i].year = 2009;
